#368 - Make possible to delete features and whole layers
- Factor detail forms out into separate classes to make the code more manageable
